Get latest CDI version using GitHub CLI
Previously, we queried the latest version of CDI with an unauthenticated
request from the command line using 'curl'. Frequently it failed with
the following error:

"API rate limit exceeded ... (But here's the good news: Authenticated
requests get a higer rate limit ..."

So here we change the way the latest version of CDI is queried. It is
done using GitHub CLI with the access token from the GitHub repository.
However, in order to keep the scripts working locally without setting an
access token, there is a hard coded version that will be used when the
CDI_VERSION variable is not set.

Signed-off-by: Arik Hadas <ahadas@redhat.com>